Terminal changes

+++ DPUWebProject/src/main/java/com/dpu/controller/TerminalController.java
@@ -3,8 +3,10 @@
  */
 package com.dpu.controller;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
+
 import org.apache.commons.beanutils.BeanUtils;
 import org.apache.log4j.Logger;
 import org.codehaus.jackson.map.ObjectMapper;
@@ -18,15 +20,13 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+
 import com.dpu.constants.Iconstants;
-import com.dpu.entity.Company;
 import com.dpu.entity.Terminal;
-import com.dpu.model.CompanyResponse;
 import com.dpu.model.Failed;
 import com.dpu.model.Success;
 import com.dpu.model.TerminalResponse;
 import com.dpu.service.TerminalService;
-import com.dpu.util.MessageProperties;
 
 /**
  * @author anuj
@@ -169,7 +169,7 @@ public class TerminalController {
 	}
 
 	@RequestMapping(value = "/{terminalid}", produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.PUT)
-	public Object update(@PathVariable("terminalid") int id, @RequestBody Terminal terminal) {
+	public Object update(@PathVariable("terminalid") BigInteger id, @RequestBody Terminal terminal) {
 		logger.info("[update] : Enter : Id : "+id);
 		Object obj = null;
 		try {
